requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-silk==4.1.0
djangorestframework==3.12.4
Faker==8.1.2
gprof2dot==2021.2.21
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.20.2
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
